fix: refactoring code structure
