+++ skyagi/src/skyagi/model.py
@@ -1,0 +1,58 @@
+from typing import Dict, Type
+
+from langchain import chat_models, embeddings, llms
+from langchain.embeddings.base import Embeddings
+from langchain.llms.base import BaseLanguageModel
+
+from skyagi.settings import EmbeddingSettings, LLMSettings
+
+# ------------------------- LLM/Chat models registry ------------------------- #
+llm_type_to_cls_dict: Dict[str, Type[BaseLanguageModel]] = {
+    "chatopenai": chat_models.ChatOpenAI,
+    "openai": llms.OpenAI,
+}
+
+# ------------------------- Embedding models registry ------------------------ #
+embedding_type_to_cls_dict: Dict[str, Type[Embeddings]] = {
+    "openaiembeddings": embeddings.OpenAIEmbeddings
+}
+
+
+# ---------------------------------------------------------------------------- #
+#                                LLM/Chat models                               #
+# ---------------------------------------------------------------------------- #
+def load_llm_from_config(config: LLMSettings) -> BaseLanguageModel:
+    """Load LLM from Config."""
+    config_dict = config.dict()
+    config_type = config_dict.pop("type")
+
+    if config_type not in llm_type_to_cls_dict:
+        raise ValueError(f"Loading {config_type} LLM not supported")
+
+    cls = llm_type_to_cls_dict[config_type]
+    return cls(**config_dict)
+
+
+def get_all_llms() -> list[str]:
+    """Get all supported LLMs"""
+    return list(llm_type_to_cls_dict.keys())
+
+
+# ---------------------------------------------------------------------------- #
+#                               Embeddings models                              #
+# ---------------------------------------------------------------------------- #
+def load_embedding_from_config(config: EmbeddingSettings) -> Embeddings:
+    """Load Embedding from Config."""
+    config_dict = config.dict()
+    config_type = config_dict.pop("type")
+
+    if config_type not in embedding_type_to_cls_dict:
+        raise ValueError(f"Loading {config_type} Embedding not supported")
+
+    cls = embedding_type_to_cls_dict[config_type]
+    return cls(**config_dict)
+
+
+def get_all_embeddings() -> list[str]:
+    """Get all supported Embeddings"""
+    return list(embedding_type_to_cls_dict.keys())
